+++ src/avdecc/mappingsHelper.hpp
@@ -20,6 +20,17 @@
 #pragma once
 
 #include "mappingMatrix.hpp"
+#include "helper.hpp"
+#include "controllerManager.hpp"
+
+#include <la/avdecc/avdecc.hpp>
+#include <la/avdecc/controller/avdeccController.hpp>
+
+#include <cstdint>
+#include <vector>
+#include <set>
+#include <utility>
+#include <QObject>
 
 namespace avdecc
 {
@@ -270,6 +281,8 @@ void processNewConnections(la::avdecc::UniqueIdentifier const entityID, la::avde
 		}
 	}
 }
+
+void showMappingsEditor(QObject* obj, la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::DescriptorType const streamPortType, la::avdecc::entity::model::StreamPortIndex const streamPortIndex, la::avdecc::entity::model::StreamIndex const streamIndex) noexcept;
 
 } // namespace mappingsHelper
 } // namespace avdecc